refactor: 폴더 구조 변경 및 리팩터링, router 적용
