+++ week4-ts/src/components/Error404.tsx
@@ -1,0 +1,3 @@
+export default function Error404() {
+    return <h1>404 Not Found</h1>
+}
