+++ week4-ts/src/types/PubInfo.ts
@@ -1,0 +1,8 @@
+export interface PubInfo {
+    readonly id?: number;
+    readonly place_url?: string;
+    readonly place_name?: string;
+    readonly phone?: string;
+    readonly distance?: string;
+    readonly adress_name?: string;
+  }
